requirements.txt
psycopg~=3.2.3
pytxc~=0.1.6
python-dotenv~=1.0.1
bods-client~=0.12.0
flit~=3.1.1
